Use SIGUSR1 to trigger save-and-exit
This is useful when editing a file in one SSH session, moving to another
PC and establishing a new SSH session to the same host, and then trying
to edit the same file in the new session (because you never remember to
use tmux/screen when you'll actually need it).

You can run `killall -USR1 nvim` and then start a new neovim session to
resume editing the document(s) in question.